Use San Francisco font
San Francisco is available in Safari as -apple-system but is not usable
by name, like other fonts. Chrome does not support -apple-system, so
we’ll use the secret, hidden name of the font for now.

This name is subject to change at any time. Hopefully we can use
`font-face: system` in the future.